+++ src/_spec/compiler-9.spec.ts
@@ -61,10 +61,114 @@ describe("compiler-9", function() {
             }
         }
     });
+    it("compiler-declare-statement-1", function() {
+        const schemas = [compile(`
+            declare var x: number;
+            declare let y: number[];
+            declare const z: \nnumber[][];
+        `)];
+
+        {
+            expect(Array.from(schemas[0].keys())).toEqual([
+                '__$$$gensym_0$$$__',
+                '__$$$gensym_1$$$__',
+                '__$$$gensym_2$$$__',
+            ]);
+        }
+        for (const schema of schemas) {
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `declare var x: number;`,
+                };
+                const ty = getType(schema, '__$$$gensym_0$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `declare let y: number[];`,
+                };
+                const ty = getType(schema, '__$$$gensym_1$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `declare const z: \nnumber[][];`,
+                };
+                const ty = getType(schema, '__$$$gensym_2$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                expect(generateTypeScriptCode(schema).trim()).toEqual(
+                    `declare var x: number;\n\n` +
+                    `declare let y: number[];\n\n` +
+                    `declare const z: \nnumber[][];`
+                );
+            }
+        }
+    });
+    it("compiler-declare-statement-2", function() {
+        const schemas = [compile(`
+            export declare var x: number;
+            export declare let y: number[];
+            export declare const z: \nnumber[][];
+        `)];
+
+        {
+            expect(Array.from(schemas[0].keys())).toEqual([
+                '__$$$gensym_0$$$__',
+                '__$$$gensym_1$$$__',
+                '__$$$gensym_2$$$__',
+            ]);
+        }
+        for (const schema of schemas) {
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `export declare var x: number;`,
+                };
+                const ty = getType(schema, '__$$$gensym_0$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `export declare let y: number[];`,
+                };
+                const ty = getType(schema, '__$$$gensym_1$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                const rhs: TypeAssertion = {
+                    kind: 'never',
+                    passThruCodeBlock: `export declare const z: \nnumber[][];`,
+                };
+                const ty = getType(schema, '__$$$gensym_2$$$__');
+                expect(ty).toEqual(rhs);
+            }
+            {
+                expect(generateTypeScriptCode(schema).trim()).toEqual(
+                    `export declare var x: number;\n\n` +
+                    `export declare let y: number[];\n\n` +
+                    `export declare const z: \nnumber[][];`
+                );
+            }
+        }
+    });
     it("compiler-directives-1", function() {
         const schemas = [compile(`
             /// @tynder-external P, Q, R
             // @tynder-external S
+
+            type X = Q;
+            interface Y { a: R; }
+        `), compile(`
+            /* @tynder-external P, Q, R */
+            /* @tynder-external
+            S
+            */
 
             type X = Q;
             interface Y { a: R; }